Pause other client as soon as one client is paused
If one player clicks quit, the other player also gets the pause dialog.
If one player continues, the other player also continues the game.
Going back to start is not implemented yet. Only pause and continue for now.
The countdown is also not synced, so it can happen for now, that one client is a second faster than the other, depending on the connection.