Redirect other player to next level
If one player confirms the level success message and continues to the next level,
the other player is also redirected there.
To keep the team data (level, score, lifes) in snyc
and to always display the right data each time
the game view is entered,
the data is now updated from the backend in the
game view model constructor.