Fix inconsistencies in game play
The game running status is not only true or false no, but indicates the triggered action.
This way it is easier to determine the needed action that should be executed, when the game running status change comes in from the socket.

Also we are waiting with continuing to the next level now until the save is done to make sure, the other client also gets the correct changes.

The team model is now updated with the newest data as soon as pause is clicked - the client that is receiving the socket connection update does the same. This way, the correct data should be loaded when the game view is reloaded on pause continue.